lab8

+++ src/lab5/WeightedEdge.java
@@ -4,10 +4,10 @@ import lab4.AbstractGraph;
 
 public class WeightedEdge extends AbstractGraph.Edge
      implements Comparable<WeightedEdge>  {
-  public double weight; // The weight on edge (u, v)
+  public int weight; // The weight on edge (u, v)
 
   /** Create a weighted edge on (u, v) */
-  public WeightedEdge(int u, int v, double weight) {
+  public WeightedEdge(int u, int v, int weight) {
     super(u, v);
     this.weight = weight;
   }
